backend/pyproject.toml
[tool.poetry]
name = "reflow-ocr-backend"
version = "0.1.0"
description = "FastAPI backend service for the Reflow OCR desktop application."
authors = ["Alex"]
license = "Apache-2.0"
readme = "README.md"
packages = [
    { include = "reflow_ocr" }
]

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.115.0"
uvicorn = { version = "^0.30.0", extras = ["standard"] }
orjson = "^3.10.0"
pydantic = "^2.9.0"
pydantic-settings = "^2.4.0"
python-multipart = "^0.0.9"
aiofiles = "^24.1.0"
numpy = "^2.1.0"
opencv-python-headless = "^4.10.0.0"
pillow = "^10.4.0"
rich = "^13.7.1"
httpx = "^0.27.0"
platformdirs = "^4.3.0"
python-docx = "^1.1.0"
reportlab = "^4.2.2"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.0"
pytest-asyncio = "^0.24.0"
ruff = "^0.6.0"
mypy = "^1.11.0"
types-requests = "^2.32.0"

[build-system]
requires = ["poetry-core>=1.8.0"]
build-backend = "poetry.core.masonry.api"
